requirements.txt
matplotlib==3.4.2
youtube_search_python==1.4.6
numpy==1.20.3
youtube_transcript_api==0.4.1
streamlit==0.83.0
pandas==1.2.4
wordcloud==1.8.1
Pillow==8.2.0
XlsxWriter==1.4.3